+++ app/src/helpers/review_state/review_state_test.rs
@@ -4,7 +4,7 @@ mod test_review_state {
 
     use crate::helpers::review_state::ReviewState;
 
-    fn approx(a: f32, b: f32) -> bool {
+    fn approx(a: f64, b: f64) -> bool {
         (a - b).abs() < 1e-6
     }
